@@ -21,43 +21,46 @@
     ],
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0-dev",
-            "dev-develop": "1.1-dev"
+            "dev-master": "1.2-dev",
+            "dev-develop": "1.3-dev"
         }
     },
     "require": {
-        "php": ">=5.3.23",
-        "zendframework/zend-eventmanager": "~2.3",
-        "zendframework/zend-json": "~2.3",
-        "zendframework/zend-loader": "~2.3",
-        "zendframework/zend-mvc": "~2.3.0",
-        "zendframework/zend-paginator": "~2.3",
-        "zendframework/zend-stdlib": "~2.3",
+        "php": ">=5.5",
+        "zendframework/zend-eventmanager": "^2.4.0",
+        "zendframework/zend-json": "^2.4.0",
+        "zendframework/zend-loader": "^2.4.0",
+        "zendframework/zend-mvc": "^2.4.0",
+        "zendframework/zend-paginator": "^2.4.0",
+        "zendframework/zend-stdlib": "^2.4.0",
         "zfcampus/zf-api-problem": "~1.0",
         "zfcampus/zf-content-negotiation": "~1.0",
-        "zfcampus/zf-hal": ">=1.0.2,<2.0.0",
-        "zfcampus/zf-mvc-auth": "~1.0",
-        "zfcampus/zf-oauth2": "~1.0"
+        "zfcampus/zf-hal": "~1.1",
+        "zfcampus/zf-mvc-auth": "~1.1"
     },
     "require-dev": {
-        "phpunit/PHPUnit": "3.7.*",
-        "squizlabs/php_codesniffer": "~2.0",
-        "zendframework/zend-console": "~2.3",
-        "zendframework/zend-escaper": "~2.3",
-        "zendframework/zend-http": "~2.3",
-        "zendframework/zend-inputfilter": "~2.3",
-        "zendframework/zend-servicemanager": "~2.3",
-        "zendframework/zend-uri": "~2.3",
-        "zendframework/zend-validator": "~2.3",
-        "zendframework/zend-view": "~2.3"
+        "phpunit/phpunit": "~4.7",
+        "squizlabs/php_codesniffer": "^2.3.1",
+        "zendframework/zend-console": "^2.4.0",
+        "zendframework/zend-escaper": "^2.4.0",
+        "zendframework/zend-http": "^2.4.0",
+        "zendframework/zend-inputfilter": "^2.4.0",
+        "zendframework/zend-servicemanager": "^2.4.0",
+        "zendframework/zend-uri": "^2.4.0",
+        "zendframework/zend-validator": "^2.4.0",
+        "zendframework/zend-view": "^2.4.0"
     },
     "autoload": {
         "psr-4": {
-            "ZF\\Rest\\": "src/",
-            "ZFTest\\Rest\\": "test/"
+            "ZF\\Rest\\": "src/"
         },
         "classmap": [
             "Module.php"
         ]
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "ZFTest\\Rest\\": "test/"
+        }
     }
 }
