@@ -7,10 +7,12 @@
 namespace ZFTest\Rest;
 
 use PHPUnit_Framework_TestCase as TestCase;
+use ReflectionMethod;
 use ReflectionObject;
 use stdClass;
 use Zend\EventManager\EventManager;
 use Zend\EventManager\SharedEventManager;
+use Zend\Http\Response;
 use Zend\InputFilter\InputFilter;
 use Zend\Mvc\Controller\PluginManager;
 use Zend\Mvc\MvcEvent;
@@ -27,6 +29,9 @@ use ZF\ContentNegotiation\ControllerPlugin\BodyParams;
 use ZF\ContentNegotiation\ParameterDataContainer;
 use ZF\Hal\Collection as HalCollection;
 use ZF\Hal\Entity as HalEntity;
+use ZF\Hal\Extractor\LinkCollectionExtractor;
+use ZF\Hal\Extractor\LinkExtractor;
+use ZF\Hal\Link\Link;
 use ZF\Hal\Plugin\Hal as HalHelper;
 use ZF\Rest\Exception;
 use ZF\Rest\Resource;
@@ -46,7 +51,7 @@ class RestControllerTest extends TestCase
         $router->addRoute('resource', $route);
         $this->event = $event = new MvcEvent();
         $event->setRouter($router);
-        $event->setRouteMatch(new RouteMatch(array()));
+        $event->setRouteMatch(new RouteMatch([]));
         $controller->setEvent($event);
         $controller->setRoute('resource');
 
@@ -65,6 +70,10 @@ class RestControllerTest extends TestCase
         $linksHelper->setUrlHelper($urlHelper);
         $linksHelper->setServerUrlHelper($serverUrlHelper);
 
+        $linkExtractor = new LinkExtractor($serverUrlHelper, $urlHelper);
+        $linkCollectionExtractor = new LinkCollectionExtractor($linkExtractor);
+        $linksHelper->setLinkCollectionExtractor($linkCollectionExtractor);
+
         $pluginManager->setService('Hal', $linksHelper);
         $linksHelper->setController($controller);
 
@@ -74,11 +83,11 @@ class RestControllerTest extends TestCase
 
     public function testReturnsErrorResponseWhenPageSizeExceedsMax()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz'),
-            array('id' => 'bar', 'bar' => 'baz'),
-            array('id' => 'baz', 'bar' => 'baz'),
-        );
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz'],
+            ['id' => 'bar', 'bar' => 'baz'],
+            ['id' => 'baz', 'bar' => 'baz'],
+        ];
         $adapter   = new ArrayPaginator($items);
         $paginator = new Paginator($adapter);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($paginator) {
@@ -88,22 +97,22 @@ class RestControllerTest extends TestCase
         $this->controller->setPageSizeParam('page_size');
         $this->controller->setMaxPageSize(2);
         $request = $this->controller->getRequest();
-        $request->setQuery(new Parameters(array(
+        $request->setQuery(new Parameters([
             'page'      => 1,
             'page_size' => 3,
-        )));
+        ]));
 
         $result = $this->controller->getList();
-        $this->assertProblemApiResult(500, "Page size is out of range, maximum page size is 2", $result);
+        $this->assertProblemApiResult(416, "Page size is out of range, maximum page size is 2", $result);
     }
 
     public function testReturnsErrorResponseWhenPageSizeBelowMin()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz'),
-            array('id' => 'bar', 'bar' => 'baz'),
-            array('id' => 'baz', 'bar' => 'baz'),
-        );
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz'],
+            ['id' => 'bar', 'bar' => 'baz'],
+            ['id' => 'baz', 'bar' => 'baz'],
+        ];
         $adapter   = new ArrayPaginator($items);
         $paginator = new Paginator($adapter);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($paginator) {
@@ -113,13 +122,27 @@ class RestControllerTest extends TestCase
         $this->controller->setPageSizeParam('page_size');
         $this->controller->setMinPageSize(2);
         $request = $this->controller->getRequest();
-        $request->setQuery(new Parameters(array(
+        $request->setQuery(new Parameters([
             'page'      => 1,
             'page_size' => 1,
-        )));
+        ]));
 
         $result = $this->controller->getList();
-        $this->assertProblemApiResult(500, "Page size is out of range, minimum page size is 2", $result);
+        $this->assertProblemApiResult(416, "Page size is out of range, minimum page size is 2", $result);
+    }
+
+    /**
+     * @group hotfix/77
+     */
+    public function testReturnsErrorResponseWhenPageNonInteger()
+    {
+        $request = $this->controller->getRequest();
+        $request->setQuery(new Parameters([
+            'page'      => '1/',
+        ]));
+
+        $result = $this->controller->getList();
+        $this->assertProblemApiResult(400, 'Page must be an integer; received "string"', $result);
     }
 
     public function assertProblemApiResult($expectedStatus, $expectedDetail, $result)
@@ -136,7 +159,7 @@ class RestControllerTest extends TestCase
             throw new Exception\CreationException('failed');
         });
 
-        $result = $this->controller->create(array());
+        $result = $this->controller->create([]);
         $this->assertProblemApiResult(500, 'failed', $result);
     }
 
@@ -145,13 +168,13 @@ class RestControllerTest extends TestCase
      *
      * @group 43
      */
-    public function testCreateDoesNotSetLocationHeaderOnMissingEntityIdentifier()
+    public function testCreateDoesNotSetLocationHeaderOnMissingSelfRelationalLink()
     {
         $this->resource->getEventManager()->attach('create', function ($e) {
-            return array('foo' => 'bar');
+            return new HalEntity(['foo' => 'bar']);
         });
 
-        $result = $this->controller->create(array());
+        $result = $this->controller->create([]);
         $this->assertInstanceOf('ZF\Hal\Entity', $result);
         $response = $this->controller->getResponse();
         $headers  = $response->getHeaders();
@@ -160,12 +183,12 @@ class RestControllerTest extends TestCase
 
     public function testCreateReturnsHalEntityOnSuccess()
     {
-        $entity = array('id' => 'foo', 'bar' => 'baz');
+        $entity = ['id' => 'foo', 'bar' => 'baz'];
         $this->resource->getEventManager()->attach('create', function ($e) use ($entity) {
             return $entity;
         });
 
-        $result = $this->controller->create(array());
+        $result = $this->controller->create([]);
         $this->assertInstanceOf('ZF\Hal\Entity', $result);
         $this->assertEquals($entity, $result->getEntity());
         return $this->controller->getResponse();
@@ -207,7 +230,7 @@ class RestControllerTest extends TestCase
             return false;
         });
 
-        $result = $this->controller->deleteList();
+        $result = $this->controller->deleteList(null);
         $this->assertProblemApiResult(422, 'delete collection', $result);
     }
 
@@ -217,7 +240,7 @@ class RestControllerTest extends TestCase
             return true;
         });
 
-        $result = $this->controller->deleteList(array(1, 2, 3));
+        $result = $this->controller->deleteList([1, 2, 3]);
         $this->assertInstanceOf('Zend\Http\Response', $result);
         $this->assertEquals(204, $result->getStatusCode());
     }
@@ -228,7 +251,7 @@ class RestControllerTest extends TestCase
             return true;
         });
 
-        $result = $this->controller->deleteList();
+        $result = $this->controller->deleteList(null);
         $this->assertInstanceOf('Zend\Http\Response', $result);
         $this->assertEquals(204, $result->getStatusCode());
     }
@@ -245,7 +268,7 @@ class RestControllerTest extends TestCase
 
     public function testReturningEntityFromGetReturnsExpectedHalEntity()
     {
-        $entity = array('id' => 'foo', 'bar' => 'baz');
+        $entity = ['id' => 'foo', 'bar' => 'baz'];
         $this->resource->getEventManager()->attach('fetch', function ($e) use ($entity) {
             return $entity;
         });
@@ -257,9 +280,9 @@ class RestControllerTest extends TestCase
 
     public function testReturnsHalCollectionForNonPaginatedList()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz')
-        );
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz']
+        ];
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($items) {
             return $items;
         });
@@ -272,11 +295,11 @@ class RestControllerTest extends TestCase
 
     public function testReturnsHalCollectionForPaginatedList()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz'),
-            array('id' => 'bar', 'bar' => 'baz'),
-            array('id' => 'baz', 'bar' => 'baz'),
-        );
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz'],
+            ['id' => 'bar', 'bar' => 'baz'],
+            ['id' => 'baz', 'bar' => 'baz'],
+        ];
         $adapter   = new ArrayPaginator($items);
         $paginator = new Paginator($adapter);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($paginator) {
@@ -285,7 +308,7 @@ class RestControllerTest extends TestCase
 
         $this->controller->setPageSize(1);
         $request = $this->controller->getRequest();
-        $request->setQuery(new Parameters(array('page' => 2)));
+        $request->setQuery(new Parameters(['page' => 2]));
 
         $result = $this->controller->getList();
         $this->assertInstanceOf('ZF\Hal\Collection', $result);
@@ -296,11 +319,11 @@ class RestControllerTest extends TestCase
 
     public function testReturnsHalCollectionForPaginatedListUsingPassedPageSizeParameter()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz'),
-            array('id' => 'bar', 'bar' => 'baz'),
-            array('id' => 'baz', 'bar' => 'baz'),
-        );
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz'],
+            ['id' => 'bar', 'bar' => 'baz'],
+            ['id' => 'baz', 'bar' => 'baz'],
+        ];
         $adapter   = new ArrayPaginator($items);
         $paginator = new Paginator($adapter);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($paginator) {
@@ -309,10 +332,10 @@ class RestControllerTest extends TestCase
 
         $this->controller->setPageSizeParam('page_size');
         $request = $this->controller->getRequest();
-        $request->setQuery(new Parameters(array(
+        $request->setQuery(new Parameters([
             'page'      => 2,
             'page_size' => 1,
-        )));
+        ]));
 
         $result = $this->controller->getList();
         $this->assertInstanceOf('ZF\Hal\Collection', $result);
@@ -332,11 +355,11 @@ class RestControllerTest extends TestCase
 
     public function testHeadReturnsListResponseWhenNoIdProvided()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz'),
-            array('id' => 'bar', 'bar' => 'baz'),
-            array('id' => 'baz', 'bar' => 'baz'),
-        );
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz'],
+            ['id' => 'bar', 'bar' => 'baz'],
+            ['id' => 'baz', 'bar' => 'baz'],
+        ];
         $adapter   = new ArrayPaginator($items);
         $paginator = new Paginator($adapter);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($paginator) {
@@ -345,7 +368,7 @@ class RestControllerTest extends TestCase
 
         $this->controller->setPageSize(1);
         $request = $this->controller->getRequest();
-        $request->setQuery(new Parameters(array('page' => 2)));
+        $request->setQuery(new Parameters(['page' => 2]));
 
         $result = $this->controller->head();
         $this->assertInstanceOf('ZF\Hal\Collection', $result);
@@ -354,7 +377,7 @@ class RestControllerTest extends TestCase
 
     public function testHeadReturnsEntityResponseWhenIdProvided()
     {
-        $entity = array('id' => 'foo', 'bar' => 'baz');
+        $entity = ['id' => 'foo', 'bar' => 'baz'];
         $this->resource->getEventManager()->attach('fetch', function ($e) use ($entity) {
             return $entity;
         });
@@ -436,13 +459,13 @@ class RestControllerTest extends TestCase
             throw new Exception\PatchException('failed');
         });
 
-        $result = $this->controller->patch('foo', array());
+        $result = $this->controller->patch('foo', []);
         $this->assertProblemApiResult(500, 'failed', $result);
     }
 
     public function testPatchReturnsHalEntityOnSuccess()
     {
-        $entity = array('id' => 'foo', 'bar' => 'baz');
+        $entity = ['id' => 'foo', 'bar' => 'baz'];
         $this->resource->getEventManager()->attach('patch', function ($e) use ($entity) {
             return $entity;
         });
@@ -458,13 +481,13 @@ class RestControllerTest extends TestCase
             throw new Exception\UpdateException('failed');
         });
 
-        $result = $this->controller->update('foo', array());
+        $result = $this->controller->update('foo', []);
         $this->assertProblemApiResult(500, 'failed', $result);
     }
 
     public function testUpdateReturnsHalEntityOnSuccess()
     {
-        $entity = array('id' => 'foo', 'bar' => 'baz');
+        $entity = ['id' => 'foo', 'bar' => 'baz'];
         $this->resource->getEventManager()->attach('update', function ($e) use ($entity) {
             return $entity;
         });
@@ -480,15 +503,15 @@ class RestControllerTest extends TestCase
             throw new Exception\UpdateException('failed');
         });
 
-        $result = $this->controller->replaceList(array());
+        $result = $this->controller->replaceList([]);
         $this->assertProblemApiResult(500, 'failed', $result);
     }
 
     public function testReplaceListReturnsHalCollectionOnSuccess()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz'),
-            array('id' => 'bar', 'bar' => 'baz'));
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz'],
+            ['id' => 'bar', 'bar' => 'baz']];
         $this->resource->getEventManager()->attach('replaceList', function ($e) use ($items) {
             return $items;
         });
@@ -524,12 +547,12 @@ class RestControllerTest extends TestCase
 
     public function testValidMethodReturningHalOrApiValueIsCastToViewModel()
     {
-        $entity = array('id' => 'foo', 'bar' => 'baz');
+        $entity = ['id' => 'foo', 'bar' => 'baz'];
         $this->resource->getEventManager()->attach('fetch', function ($e) use ($entity) {
             return $entity;
         });
 
-        $this->controller->setEntityHttpMethods(array('GET'));
+        $this->controller->setEntityHttpMethods(['GET']);
 
         $request = $this->controller->getRequest();
         $request->setMethod('GET');
@@ -542,12 +565,12 @@ class RestControllerTest extends TestCase
 
     public function testValidMethodReturningHalOrApiValueCastsReturnToContentNegotiationViewModel()
     {
-        $entity = array('id' => 'foo', 'bar' => 'baz');
+        $entity = ['id' => 'foo', 'bar' => 'baz'];
         $this->resource->getEventManager()->attach('fetch', function ($e) use ($entity) {
             return $entity;
         });
 
-        $this->controller->setEntityHttpMethods(array('GET'));
+        $this->controller->setEntityHttpMethods(['GET']);
 
         $request = $this->controller->getRequest();
         $request->setMethod('GET');
@@ -573,15 +596,15 @@ class RestControllerTest extends TestCase
             $test->flag = true;
         });
 
-        $events->trigger('test', $controller, array());
+        $events->trigger('test', $controller, []);
         $this->assertTrue($test->flag);
     }
 
     public function testHalCollectionUsesControllerCollectionName()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz')
-        );
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz']
+        ];
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($items) {
             return $items;
         });
@@ -595,7 +618,7 @@ class RestControllerTest extends TestCase
 
     public function testCreateUsesHalEntityReturnedByResource()
     {
-        $data     = array('id' => 'foo', 'data' => 'bar');
+        $data     = ['id' => 'foo', 'data' => 'bar'];
         $entity   = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('create', function ($e) use ($entity) {
             return $entity;
@@ -607,7 +630,7 @@ class RestControllerTest extends TestCase
 
     public function testGetUsesHalEntityReturnedByResource()
     {
-        $data     = array('id' => 'foo', 'data' => 'bar');
+        $data     = ['id' => 'foo', 'data' => 'bar'];
         $entity   = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('fetch', function ($e) use ($entity) {
             return $entity;
@@ -619,7 +642,7 @@ class RestControllerTest extends TestCase
 
     public function testGetListUsesHalCollectionReturnedByResource()
     {
-        $collection = new HalCollection(array());
+        $collection = new HalCollection([]);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($collection) {
             return $collection;
         });
@@ -630,7 +653,7 @@ class RestControllerTest extends TestCase
 
     public function testPatchUsesHalEntityReturnedByResource()
     {
-        $data     = array('id' => 'foo', 'data' => 'bar');
+        $data     = ['id' => 'foo', 'data' => 'bar'];
         $entity   = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('patch', function ($e) use ($entity) {
             return $entity;
@@ -642,7 +665,7 @@ class RestControllerTest extends TestCase
 
     public function testUpdateUsesHalEntityReturnedByResource()
     {
-        $data     = array('id' => 'foo', 'data' => 'bar');
+        $data     = ['id' => 'foo', 'data' => 'bar'];
         $entity   = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('update', function ($e) use ($entity) {
             return $entity;
@@ -654,24 +677,24 @@ class RestControllerTest extends TestCase
 
     public function testReplaceListUsesHalCollectionReturnedByResource()
     {
-        $collection = new HalCollection(array());
+        $collection = new HalCollection([]);
         $this->resource->getEventManager()->attach('replaceList', function ($e) use ($collection) {
             return $collection;
         });
 
-        $result = $this->controller->replaceList(array());
+        $result = $this->controller->replaceList([]);
         $this->assertSame($collection, $result);
     }
 
     public function testCreateTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'       => false,
             'pre_data'  => false,
             'post'      => false,
             'post_data' => false,
             'entity'    => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('create.pre', function ($e) use ($test) {
             $test->pre      = true;
@@ -683,7 +706,7 @@ class RestControllerTest extends TestCase
             $test->entity = $e->getParam('entity');
         });
 
-        $data   = array('id' => 'foo', 'data' => 'bar');
+        $data   = ['id' => 'foo', 'data' => 'bar'];
         $entity = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('create', function ($e) use ($entity) {
             return $entity;
@@ -699,12 +722,12 @@ class RestControllerTest extends TestCase
 
     public function testDeleteTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'       => false,
             'pre_id'  => false,
             'post'      => false,
             'post_id' => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('delete.pre', function ($e) use ($test) {
             $test->pre      = true;
@@ -728,10 +751,10 @@ class RestControllerTest extends TestCase
 
     public function testDeleteListTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'       => false,
             'post'      => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('deleteList.pre', function ($e) use ($test) {
             $test->pre      = true;
@@ -744,20 +767,20 @@ class RestControllerTest extends TestCase
             return true;
         });
 
-        $result = $this->controller->deleteList();
+        $result = $this->controller->deleteList(null);
         $this->assertTrue($test->pre);
         $this->assertTrue($test->post);
     }
 
     public function testGetTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'       => false,
             'pre_id'    => false,
             'post'      => false,
             'post_id'   => false,
             'entity'    => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('get.pre', function ($e) use ($test) {
             $test->pre    = true;
@@ -769,7 +792,7 @@ class RestControllerTest extends TestCase
             $test->entity = $e->getParam('entity');
         });
 
-        $data   = array('id' => 'foo', 'data' => 'bar');
+        $data   = ['id' => 'foo', 'data' => 'bar'];
         $entity = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('fetch', function ($e) use ($entity) {
             return $entity;
@@ -785,15 +808,15 @@ class RestControllerTest extends TestCase
 
     public function testOptionsTriggersPreAndPostEventsForCollection()
     {
-        $methods = array('GET', 'POST');
+        $methods = ['GET', 'POST'];
         $this->controller->setCollectionHttpMethods($methods);
 
-        $test = (object) array(
+        $test = (object) [
             'pre'          => false,
             'post'         => false,
             'pre_options'  => false,
             'post_options' => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('options.pre', function ($e) use ($test) {
             $test->pre = true;
@@ -813,15 +836,15 @@ class RestControllerTest extends TestCase
 
     public function testOptionsTriggersPreAndPostEventsForEntity()
     {
-        $methods = array('GET', 'PUT', 'PATCH');
+        $methods = ['GET', 'PUT', 'PATCH'];
         $this->controller->setEntityHttpMethods($methods);
 
-        $test = (object) array(
+        $test = (object) [
             'pre'          => false,
             'post'         => false,
             'pre_options'  => false,
             'post_options' => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('options.pre', function ($e) use ($test) {
             $test->pre = true;
@@ -843,11 +866,11 @@ class RestControllerTest extends TestCase
 
     public function testGetListTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'        => false,
             'post'       => false,
             'collection' => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('getList.pre', function ($e) use ($test) {
             $test->pre    = true;
@@ -857,7 +880,7 @@ class RestControllerTest extends TestCase
             $test->collection = $e->getParam('collection');
         });
 
-        $collection = new HalCollection(array());
+        $collection = new HalCollection([]);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($collection) {
             return $collection;
         });
@@ -874,7 +897,7 @@ class RestControllerTest extends TestCase
 
     public function testPatchTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'       => false,
             'pre_id'    => false,
             'pre_data'  => false,
@@ -882,7 +905,7 @@ class RestControllerTest extends TestCase
             'post_id'   => false,
             'post_data' => false,
             'entity'    => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('patch.pre', function ($e) use ($test) {
             $test->pre      = true;
@@ -896,7 +919,7 @@ class RestControllerTest extends TestCase
             $test->entity    = $e->getParam('entity');
         });
 
-        $data     = array('id' => 'foo', 'data' => 'bar');
+        $data     = ['id' => 'foo', 'data' => 'bar'];
         $entity   = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('patch', function ($e) use ($entity) {
             return $entity;
@@ -914,7 +937,7 @@ class RestControllerTest extends TestCase
 
     public function testUpdateTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'       => false,
             'pre_id'    => false,
             'pre_data'  => false,
@@ -922,7 +945,7 @@ class RestControllerTest extends TestCase
             'post_id'   => false,
             'post_data' => false,
             'entity'    => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('update.pre', function ($e) use ($test) {
             $test->pre      = true;
@@ -936,7 +959,7 @@ class RestControllerTest extends TestCase
             $test->entity    = $e->getParam('entity');
         });
 
-        $data     = array('id' => 'foo', 'data' => 'bar');
+        $data     = ['id' => 'foo', 'data' => 'bar'];
         $entity   = new HalEntity($data, 'foo', 'resource');
         $this->resource->getEventManager()->attach('update', function ($e) use ($entity) {
             return $entity;
@@ -954,13 +977,13 @@ class RestControllerTest extends TestCase
 
     public function testReplaceListTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'        => false,
             'pre_data'   => false,
             'post'       => false,
             'post_data'  => false,
             'collection' => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('replaceList.pre', function ($e) use ($test) {
             $test->pre      = true;
@@ -972,7 +995,7 @@ class RestControllerTest extends TestCase
             $test->collection = $e->getParam('collection');
         });
 
-        $data       = array('foo' => array('id' => 'bar'));
+        $data       = ['foo' => ['id' => 'bar']];
         $collection = new HalCollection($data);
         $this->resource->getEventManager()->attach('replaceList', function ($e) use ($collection) {
             return $collection;
@@ -1020,20 +1043,20 @@ class RestControllerTest extends TestCase
 
     public function eventsProducingApiProblems()
     {
-        return array(
-            'delete' => array(
+        return [
+            'delete' => [
                 'delete', 'delete', 'foo',
-            ),
-            'deleteList' => array(
+            ],
+            'deleteList' => [
                 'deleteList', 'deleteList', null,
-            ),
-            'get' => array(
+            ],
+            'get' => [
                 'fetch', 'get', 'foo',
-            ),
-            'getList' => array(
+            ],
+            'getList' => [
                 'fetchAll', 'getList', null,
-            ),
-        );
+            ],
+        ];
     }
 
     /**
@@ -1092,13 +1115,13 @@ class RestControllerTest extends TestCase
             'Validation error',
             null,
             null,
-            array('email' => 'Invalid email address provided')
+            ['email' => 'Invalid email address provided']
         );
         $this->resource->getEventManager()->attach('create', function ($e) use ($problem) {
             return $problem;
         });
 
-        $result = $this->controller->create(array());
+        $result = $this->controller->create([]);
         $this->assertSame($problem, $result);
     }
 
@@ -1112,7 +1135,7 @@ class RestControllerTest extends TestCase
             'Invalid identifier',
             null,
             null,
-            array('delete' => 'Invalid identifier provided')
+            ['delete' => 'Invalid identifier provided']
         );
         $this->resource->getEventManager()->attach('delete', function ($e) use ($problem) {
             return $problem;
@@ -1127,12 +1150,12 @@ class RestControllerTest extends TestCase
      */
     public function testDeleteListAllowsReturningApiProblemFromResource()
     {
-        $problem = new ApiProblem(400, 'Invalid list', null, null, array('delete' => 'Invalid collection'));
+        $problem = new ApiProblem(400, 'Invalid list', null, null, ['delete' => 'Invalid collection']);
         $this->resource->getEventManager()->attach('deleteList', function ($e) use ($problem) {
             return $problem;
         });
 
-        $result = $this->controller->deleteList();
+        $result = $this->controller->deleteList(null);
         $this->assertSame($problem, $result);
     }
 
@@ -1141,7 +1164,7 @@ class RestControllerTest extends TestCase
      */
     public function testGetAllowsReturningApiProblemFromResource()
     {
-        $problem = new ApiProblem(400, 'Invalid identifier', null, null, array('get' => 'Invalid identifier provided'));
+        $problem = new ApiProblem(400, 'Invalid identifier', null, null, ['get' => 'Invalid identifier provided']);
         $this->resource->getEventManager()->attach('fetch', function ($e) use ($problem) {
             return $problem;
         });
@@ -1155,7 +1178,7 @@ class RestControllerTest extends TestCase
      */
     public function testGetListAllowsReturningApiProblemFromResource()
     {
-        $problem = new ApiProblem(400, 'Invalid collection', null, null, array('fetchAll' => 'Invalid collection'));
+        $problem = new ApiProblem(400, 'Invalid collection', null, null, ['fetchAll' => 'Invalid collection']);
         $this->resource->getEventManager()->attach('fetchAll', function ($e) use ($problem) {
             return $problem;
         });
@@ -1174,13 +1197,13 @@ class RestControllerTest extends TestCase
             'Validation error',
             null,
             null,
-            array('email' => 'Invalid email address provided')
+            ['email' => 'Invalid email address provided']
         );
         $this->resource->getEventManager()->attach('patch', function ($e) use ($problem) {
             return $problem;
         });
 
-        $result = $this->controller->patch('foo', array());
+        $result = $this->controller->patch('foo', []);
         $this->assertSame($problem, $result);
     }
 
@@ -1194,13 +1217,13 @@ class RestControllerTest extends TestCase
             'Validation error',
             null,
             null,
-            array('email' => 'Invalid email address provided')
+            ['email' => 'Invalid email address provided']
         );
         $this->resource->getEventManager()->attach('update', function ($e) use ($problem) {
             return $problem;
         });
 
-        $result = $this->controller->update('foo', array());
+        $result = $this->controller->update('foo', []);
         $this->assertSame($problem, $result);
     }
 
@@ -1214,13 +1237,13 @@ class RestControllerTest extends TestCase
             'Validation error',
             null,
             null,
-            array('email' => 'Invalid email address provided')
+            ['email' => 'Invalid email address provided']
         );
         $this->resource->getEventManager()->attach('replaceList', function ($e) use ($problem) {
             return $problem;
         });
 
-        $result = $this->controller->replaceList(array());
+        $result = $this->controller->replaceList([]);
         $this->assertSame($problem, $result);
     }
 
@@ -1230,15 +1253,16 @@ class RestControllerTest extends TestCase
             throw new Exception\UpdateException('failed');
         });
 
-        $result = $this->controller->patchList(array());
+        $result = $this->controller->patchList([]);
         $this->assertProblemApiResult(500, 'failed', $result);
     }
 
     public function testPatchListReturnsHalCollectionOnSuccess()
     {
-        $items = array(
-            array('id' => 'foo', 'bar' => 'baz'),
-            array('id' => 'bar', 'bar' => 'baz'));
+        $items = [
+            ['id' => 'foo', 'bar' => 'baz'],
+            ['id' => 'bar', 'bar' => 'baz'],
+        ];
         $this->resource->getEventManager()->attach('patchList', function ($e) use ($items) {
             return $items;
         });
@@ -1259,24 +1283,24 @@ class RestControllerTest extends TestCase
 
     public function testPatchListUsesHalCollectionReturnedByResource()
     {
-        $collection = new HalCollection(array());
+        $collection = new HalCollection([]);
         $this->resource->getEventManager()->attach('patchList', function ($e) use ($collection) {
             return $collection;
         });
 
-        $result = $this->controller->patchList(array());
+        $result = $this->controller->patchList([]);
         $this->assertSame($collection, $result);
     }
 
     public function testPatchListTriggersPreAndPostEvents()
     {
-        $test = (object) array(
+        $test = (object) [
             'pre'        => false,
             'pre_data'   => false,
             'post'       => false,
             'post_data'  => false,
             'collection' => false,
-        );
+        ];
 
         $this->controller->getEventManager()->attach('patchList.pre', function ($e) use ($test) {
             $test->pre      = true;
@@ -1288,7 +1312,7 @@ class RestControllerTest extends TestCase
             $test->collection = $e->getParam('collection');
         });
 
-        $data       = array('foo' => array('id' => 'bar'));
+        $data       = ['foo' => ['id' => 'bar']];
         $collection = new HalCollection($data);
         $this->resource->getEventManager()->attach('patchList', function ($e) use ($collection) {
             return $collection;
@@ -1312,83 +1336,83 @@ class RestControllerTest extends TestCase
             'Validation error',
             null,
             null,
-            array('email' => 'Invalid email address provided')
+            ['email' => 'Invalid email address provided']
         );
         $this->resource->getEventManager()->attach('patchList', function ($e) use ($problem) {
             return $problem;
         });
 
-        $result = $this->controller->patchList(array());
+        $result = $this->controller->patchList([]);
         $this->assertSame($problem, $result);
     }
 
     public function validResourcePayloads()
     {
-        return array(
-            'GET_collection' => array(
+        return [
+            'GET_collection' => [
                 'GET',
                 'fetchAll',
                 null,
                 null,
-                array(),
-            ),
-            'GET_item' => array(
+                [],
+            ],
+            'GET_item' => [
                 'GET',
                 'fetch',
                 'foo',
                 null,
-                array('id' => 'foo', 'bar' => 'baz'),
-            ),
-            'POST' => array(
+                ['id' => 'foo', 'bar' => 'baz'],
+            ],
+            'POST' => [
                 'POST',
                 'create',
                 null,
-                array('bar' => 'baz'),
-                array('id' => 'foo', 'bar' => 'baz'),
-            ),
-            'PUT_collection' => array(
+                ['bar' => 'baz'],
+                ['id' => 'foo', 'bar' => 'baz'],
+            ],
+            'PUT_collection' => [
                 'PUT',
                 'replaceList',
                 null,
-                array(array('id' => 'foo', 'bar' => 'bat')),
-                array(array('id' => 'foo', 'bar' => 'bat')),
-            ),
-            'PUT_item' => array(
+                [['id' => 'foo', 'bar' => 'bat']],
+                [['id' => 'foo', 'bar' => 'bat']],
+            ],
+            'PUT_item' => [
                 'PUT',
                 'update',
                 'foo',
-                array('bar' => 'bat'),
-                array('id' => 'foo', 'bar' => 'bat'),
-            ),
-            'PATCH_collection' => array(
+                ['bar' => 'bat'],
+                ['id' => 'foo', 'bar' => 'bat'],
+            ],
+            'PATCH_collection' => [
                 'PATCH',
                 'patchList',
                 null,
-                array('foo' => array('bar' => 'bat')),
-                array(array('id' => 'foo', 'bar' => 'bat')),
-            ),
-            'PATCH_item' => array(
+                ['foo' => ['bar' => 'bat']],
+                [['id' => 'foo', 'bar' => 'bat']],
+            ],
+            'PATCH_item' => [
                 'PATCH',
                 'patch',
                 'foo',
-                array('bar' => 'bat'),
-                array('id' => 'foo', 'bar' => 'bat'),
-            ),
-            'DELETE_collection' => array(
+                ['bar' => 'bat'],
+                ['id' => 'foo', 'bar' => 'bat'],
+            ],
+            'DELETE_collection' => [
                 'DELETE',
                 'deleteList',
                 null,
                 null,
                 true,
-            ),
-            'DELETE_item' => array(
+            ],
+            'DELETE_item' => [
                 'DELETE',
                 'delete',
                 'foo',
                 null,
                 true,
-            ),
-        );
+            ],
+        ];
     }
 
     /**
@@ -1407,15 +1431,15 @@ class RestControllerTest extends TestCase
             return $resource;
         });
 
-        $this->controller->setCollectionHttpMethods(array('GET', 'POST', 'PUT', 'PATCH', 'DELETE'));
-        $this->controller->setEntityHttpMethods(array('GET', 'PUT', 'PATCH', 'DELETE'));
+        $this->controller->setCollectionHttpMethods(['GET', 'POST', 'PUT', 'PATCH', 'DELETE']);
+        $this->controller->setEntityHttpMethods(['GET', 'PUT', 'PATCH', 'DELETE']);
 
         $request = $this->controller->getRequest();
         $request->setMethod($method);
         $this->event->setRequest($request);
 
         $container = new ParameterDataContainer();
-        $container->setBodyParams($data);
+        $container->setBodyParams((null === $data) ? [] : $data);
         $this->event->setParam('ZFContentNegotiationParameterData', $container);
 
         if ($id) {
@@ -1451,16 +1475,16 @@ class RestControllerTest extends TestCase
         $r = new ReflectionObject($resource);
         $m = $r->getMethod('prepareEvent');
         $m->setAccessible(true);
-        $event = $m->invoke($resource, 'fetch', array());
+        $event = $m->invoke($resource, 'fetch', []);
         $this->assertSame($request, $event->getRequest());
     }
 
     public function entitiesReturnedForCollections()
     {
-        return array(
-            'with-identifier' => array((object) array('id' => 'foo', 'foo' => 'bar')),
-            'no-identifier'   => array((object) array('foo' => 'bar')),
-        );
+        return [
+            'with-identifier' => [(object) ['id' => 'foo', 'foo' => 'bar']],
+            'no-identifier'   => [(object) ['foo' => 'bar']],
+        ];
     }
 
     /**
@@ -1476,5 +1500,235 @@ class RestControllerTest extends TestCase
         $result = $this->controller->getList();
         $this->assertInstanceOf('ZF\Hal\Entity', $result);
         $this->assertSame($entity, $result->getEntity());
+    }
+
+    public function methods()
+    {
+        return [
+            'get-list'    => ['getList', 'fetchAll', [null]],
+            'get'         => ['get', 'fetch', [1]],
+            'post'        => ['create', 'create', [[]]],
+            'put-list'    => ['replaceList', 'replaceList', [[]]],
+            'put'         => ['update', 'update', [1, []]],
+            'patch-list'  => ['patchList', 'patchList', [[]]],
+            'patch'       => ['patch', 'patch', [1, []]],
+            'delete-list' => ['deleteList', 'deleteList', [[]]],
+            'delete'      => ['delete', 'delete', [1]],
+        ];
+    }
+
+    /**
+     * @group 68
+     * @dataProvider methods
+     */
+    public function testAllowsReturningResponsesReturnedFromResources($method, $event, $argv)
+    {
+        $response = new Response();
+        $response->setStatusCode(418);
+
+        $events = $this->resource->getEventManager();
+        $events->attach($event, function ($e) use ($response) {
+            return $response;
+        });
+
+        $result = call_user_func_array([$this->controller, $method], $argv);
+        $this->assertSame($response, $result);
+    }
+
+    /**
+     * @group 74
+     */
+    public function testNonArrayToReplaceListReturnsApiProblem()
+    {
+        $response = $this->controller->replaceList(new stdClass());
+        $this->assertInstanceOf('ZF\ApiProblem\ApiProblem', $response);
+        $details = $response->toArray();
+        $this->assertEquals(400, $details['status']);
+    }
+
+    /**
+     * @group 79
+     */
+    public function testAllowsReturningHalCollectionFromCreate()
+    {
+        $collection = [
+            [
+                'id'  => 1,
+                'foo' => 'bar',
+            ],
+            [
+                'id'  => 2,
+                'foo' => 'bar',
+            ],
+            [
+                'id'  => 3,
+                'foo' => 'bar',
+            ],
+        ];
+        $halCollection = new HalCollection($collection);
+
+        $resource = $this->getMockBuilder('ZF\Rest\Resource')->getMock();
+        $resource
+            ->expects($this->once())
+            ->method('create')
+            ->with($this->anything())
+            ->willReturn($halCollection);
+
+        $this->controller->setResource($resource);
+        $this->controller->setCollectionName('foo');
+        $this->controller->setPageSize(5);
+
+        $request = $this->controller->getRequest();
+        $request->getQuery()->set('page', 3);
+
+        $result = $this->controller->create([]);
+
+        $this->assertSame($halCollection, $result);
+        $this->assertEquals('resource', $halCollection->getCollectionRoute());
+        $this->assertEquals('resource', $halCollection->getEntityRoute());
+        $this->assertEquals($this->controller->getRouteIdentifierName(), $halCollection->getRouteIdentifierName());
+        $this->assertEquals('foo', $halCollection->getCollectionName());
+        $this->assertEquals(5, $halCollection->getPageSize());
+        $this->assertEquals(3, $halCollection->getPage());
+        $this->assertTrue($halCollection->getLinks()->has('self'));
+    }
+
+    /**
+     * @group 79
+     */
+    public function testAllowsReturningHalEntityFromCreate()
+    {
+        $entity = [
+            'id'  => 1,
+            'foo' => 'bar',
+        ];
+        $halEntity = new HalEntity($entity, 1);
+
+        $resource = $this->getMockBuilder('ZF\Rest\Resource')->getMock();
+        $resource
+            ->expects($this->once())
+            ->method('create')
+            ->with($this->anything())
+            ->willReturn($halEntity);
+
+        $this->controller->setResource($resource);
+
+        $result = $this->controller->create([]);
+
+        $this->assertSame($halEntity, $result);
+        $this->assertTrue($halEntity->getLinks()->has('self'));
+    }
+
+    /**
+     * @group 79
+     */
+    public function testCreateHalEntityInjectsExistingEntityWithSelfRelationalLinkIfNotPresent()
+    {
+        $entity = [
+            'id'  => 1,
+            'foo' => 'bar',
+        ];
+        $halEntity = new HalEntity($entity, 1);
+
+        $r = new ReflectionMethod($this->controller, 'createHalEntity');
+        $r->setAccessible(true);
+
+        $result = $r->invoke($this->controller, $halEntity);
+        $this->assertSame($result, $halEntity);
+        $this->assertTrue($result->getLinks()->has('self'));
+    }
+
+    /**
+     * @group 79
+     */
+    public function testCreateHalEntityDoesNotInjectExistingEntityWithSelfRelationalLinkIfAlreadyPresent()
+    {
+        $entity = [
+            'id'  => 1,
+            'foo' => 'bar',
+        ];
+        $halEntity = new HalEntity($entity, 1);
+        $self = Link::factory([
+            'rel' => 'self',
+            'url' => 'http://example.com/foo/1',
+        ]);
+        $halEntity->getLinks()->add($self);
+
+        $r = new ReflectionMethod($this->controller, 'createHalEntity');
+        $r->setAccessible(true);
+
+        $result = $r->invoke($this->controller, $halEntity);
+        $this->assertSame($result, $halEntity);
+        $this->assertTrue($result->getLinks()->has('self'));
+        $this->assertSame($self, $result->getLinks()->get('self'));
+    }
+
+    /**
+     * @group 79
+     */
+    public function testCreateHalCollectionInjectsExistingCollectionWithSelfRelationalLinkIfNotPresent()
+    {
+        $collection = [
+            [
+                'id'  => 1,
+                'foo' => 'bar',
+            ],
+            [
+                'id'  => 2,
+                'foo' => 'bar',
+            ],
+            [
+                'id'  => 3,
+                'foo' => 'bar',
+            ],
+        ];
+        $halCollection = new HalCollection($collection);
+
+        $r = new ReflectionMethod($this->controller, 'createHalCollection');
+        $r->setAccessible(true);
+
+        $result = $r->invoke($this->controller, $halCollection);
+        $this->assertSame($halCollection, $result);
+        $this->assertTrue($result->getLinks()->has('self'));
+    }
+
+    /**
+     * @group 79
+     */
+    public function testCreateHalCollectionInjectsExistingCollectionWithMetadataIfMissing()
+    {
+        $collection = [
+            [
+                'id'  => 1,
+                'foo' => 'bar',
+            ],
+            [
+                'id'  => 2,
+                'foo' => 'bar',
+            ],
+            [
+                'id'  => 3,
+                'foo' => 'bar',
+            ],
+        ];
+        $halCollection = new HalCollection($collection);
+
+        $this->controller->setCollectionName('foo');
+        $this->controller->setPageSize(5);
+
+        $request = $this->controller->getRequest();
+        $request->getQuery()->set('page', 3);
+
+        $r = new ReflectionMethod($this->controller, 'createHalCollection');
+        $r->setAccessible(true);
+
+        $result = $r->invoke($this->controller, $halCollection);
+        $this->assertSame($halCollection, $result);
+        $this->assertEquals('resource', $halCollection->getCollectionRoute());
+        $this->assertEquals('resource', $halCollection->getEntityRoute());
+        $this->assertEquals($this->controller->getRouteIdentifierName(), $halCollection->getRouteIdentifierName());
+        $this->assertEquals('foo', $halCollection->getCollectionName());
+        $this->assertEquals(5, $halCollection->getPageSize());
+        $this->assertEquals(3, $halCollection->getPage());
     }
 }
